Move karma-changes resolvers to server-side-only location
