+++ packages/lesswrong/components/spuriousChange.js
@@ -1,0 +1,1 @@
+const someVar=3042
