+++ packages/lesswrong/lib/collections/users/karmaChangesGraphQL.js
@@ -1,6 +1,5 @@
 import Users from "meteor/vulcan:users";
-import { getKarmaChanges, getKarmaChangeDateRange, getKarmaChangeNextBatchDate } from "../../karmaChanges.js";
-import { addGraphQLSchema, addGraphQLResolvers, getSetting } from 'meteor/vulcan:core';
+import { addGraphQLSchema, addGraphQLResolvers } from 'meteor/vulcan:core';
 
 addGraphQLSchema(`
   type PostKarmaChange {
@@ -44,48 +43,11 @@ addGraphQLResolvers({
 
 Users.addField([
   {
-    fieldName: 'karmaChanges',
+    fieldName: "karmaChanges",
     fieldSchema: {
       viewableBy: Users.owns,
-      type: 'KarmaChanges',
+      type: "KarmaChanges",
       optional: true,
-      resolveAs: {
-        arguments: 'startDate: Date, endDate: Date',
-        type: 'KarmaChanges',
-        resolver: async (document, {startDate,endDate}, {currentUser}) => {
-          if (!currentUser)
-            return null;
-          
-          // Grab new current user, because the current user gets set at the beginning of the request, which
-          // is out of date in this case, because we are depending on recent mutations being reflected on the current user
-          const newCurrentUser = await Users.findOne(currentUser._id)
-          
-          const settings = newCurrentUser.karmaChangeNotifierSettings
-          const now = new Date();
-          
-          // If date range isn't specified, infer it from user settings
-          if (!startDate || !endDate) {
-            // If the user has karmaChanges disabled, don't return anything
-            if (settings.updateFrequency === "disabled") return null
-            const lastOpened = newCurrentUser.karmaChangeLastOpened;
-            const lastBatchStart = newCurrentUser.karmaChangeBatchStart;
-            
-            const {start, end} = getKarmaChangeDateRange({settings, lastOpened, lastBatchStart, now}) 
-            startDate = start;
-            endDate = end;
-          }
-          
-          const nextBatchDate = getKarmaChangeNextBatchDate({settings, now});
-          
-          const alignmentForum = getSetting('forumType') === 'AlignmentForum';
-          return getKarmaChanges({
-            user: document,
-            startDate, endDate,
-            nextBatchDate,
-            af: alignmentForum,
-          });
-        },
-      },
     },
-  },
+  }
 ]);
